requirements.txt
streamlit
pandas
requests
beautifulsoup4
openai>=1.0.0
google-generativeai
python-dotenv
